+++ algo/SGL_III_c.py
@@ -1,12 +1,12 @@
 import numpy as np
-from nptyping import NDArray, Int64
+from nptyping import NDArray
 from objective import Objective
 from typing import Tuple
 import utils
 
 
 def SGL_III_c(rng: np.random.Generator, f: Objective,
-              r: int, eps: float = None) -> Tuple[NDArray[Int64], float]:
+              r: int, eps: float = None) -> Tuple[NDArray[int], float]:
     """
     Randomized algorithm for integer-lattice submodular maximization of monotone functions with cardinality
     constraints in linear time.
