+++ utils/powerset.py
@@ -1,11 +1,11 @@
 import numpy as np
 import itertools
 from objective import Objective
-from nptyping import NDArray, Int64
+from nptyping import NDArray
 from typing import Iterator
 
 
-def powerset(f: Objective) -> Iterator[NDArray[Int64]]:
+def powerset(f: Objective) -> Iterator[NDArray[int]]:
     """
     Inumerate b^n possible vectors in the integer lattice.
     :param f: integer-lattice submodular function objective
